docker/requirements.txt
pandas==1.3.5
statsmodels==0.12.0
datalab==1.2.0
google-cloud-storage
pandas-gbq
